+++ coreblocks/structs_common/rs.py
@@ -4,7 +4,7 @@ from amaranth import *
 from amaranth.lib.coding import PriorityEncoder
 from transactron import Method, def_method, TModule
 from coreblocks.params import RSLayouts, GenParams, OpType
-from transactron.core import RecordDict
+from transactron.utils import RecordDict
 from transactron.utils.transactron_helpers import make_layout
 
 __all__ = ["RS"]
