+++ test/transactions/test_transactions.py
@@ -17,10 +17,10 @@ from transactron.utils import Scheduler
 
 from transactron.core import (
     Priority,
-    TransactionScheduler,
     trivial_roundrobin_cc_scheduler,
     eager_deterministic_cc_scheduler,
 )
+from transactron.core.manager import TransactionScheduler
 from transactron.utils.dependencies import DependencyContext
