+++ transactron/core/schedulers.py
@@ -7,8 +7,9 @@ if TYPE_CHECKING:
 
 __all__ = ["eager_deterministic_cc_scheduler", "trivial_roundrobin_cc_scheduler"]
 
+
 def eager_deterministic_cc_scheduler(
-    method_map: 'MethodMap', gr: 'TransactionGraph', cc: 'TransactionGraphCC', porder: 'PriorityOrder'
+    method_map: "MethodMap", gr: "TransactionGraph", cc: "TransactionGraphCC", porder: "PriorityOrder"
 ) -> Module:
     """eager_deterministic_cc_scheduler
 
@@ -44,7 +45,7 @@ def eager_deterministic_cc_scheduler(
 
 
 def trivial_roundrobin_cc_scheduler(
-    method_map: 'MethodMap', gr: 'TransactionGraph', cc: 'TransactionGraphCC', porder: 'PriorityOrder'
+    method_map: "MethodMap", gr: "TransactionGraph", cc: "TransactionGraphCC", porder: "PriorityOrder"
 ) -> Module:
     """trivial_roundrobin_cc_scheduler
