+++ transactron/core/sugar.py
@@ -3,70 +3,16 @@ from typing import TYPE_CHECKING, Optional, Callable
 from transactron.utils import *
 from transactron.utils.assign import AssignArg
 
-from .manager import TransactionManager
-from .transaction_base import TransactionManagerKey
-
 if TYPE_CHECKING:
     from .tmodule import TModule
     from .method import Method
 
-__all__ = ["TransactionModule","def_method"]
-
-class TransactionModule(Elaboratable):
-    """
-    `TransactionModule` is used as wrapper on `Elaboratable` classes,
-    which adds support for transactions. It creates a
-    `TransactionManager` which will handle transaction scheduling
-    and can be used in definition of `Method`\\s and `Transaction`\\s.
-    The `TransactionManager` is stored in a `DependencyManager`.
-    """
-
-    def __init__(
-        self,
-        elaboratable: HasElaborate,
-        dependency_manager: Optional[DependencyManager] = None,
-        transaction_manager: Optional[TransactionManager] = None,
-    ):
-        """
-        Parameters
-        ----------
-        elaboratable: HasElaborate
-            The `Elaboratable` which should be wrapped to add support for
-            transactions and methods.
-        dependency_manager: DependencyManager, optional
-            The `DependencyManager` to use inside the transaction module.
-            If omitted, a new one is created.
-        transaction_manager: TransactionManager, optional
-            The `TransactionManager` to use inside the transaction module.
-            If omitted, a new one is created.
-        """
-        if transaction_manager is None:
-            transaction_manager = TransactionManager()
-        if dependency_manager is None:
-            dependency_manager = DependencyManager()
-        self.manager = dependency_manager
-        self.manager.add_dependency(TransactionManagerKey(), transaction_manager)
-        self.elaboratable = elaboratable
-
-    def context(self) -> DependencyContext:
-        return DependencyContext(self.manager)
-
-    def elaborate(self, platform):
-        with silence_mustuse(self.manager.get_dependency(TransactionManagerKey())):
-            with self.context():
-                elaboratable = Fragment.get(self.elaboratable, platform)
-
-        m = Module()
-
-        m.submodules.main_module = elaboratable
-        m.submodules.transactionManager = self.manager.get_dependency(TransactionManagerKey())
-
-        return m
+__all__ = ["def_method"]
 
 
 def def_method(
-    m: 'TModule',
-    method: 'Method',
+    m: "TModule",
+    method: "Method",
     ready: ValueLike = C(1),
     validate_arguments: Optional[Callable[..., ValueLike]] = None,
 ):
