+++ transactron/lib/connectors.py
@@ -3,7 +3,7 @@ import amaranth.lib.fifo
 
 from transactron.utils.transactron_helpers import from_method_layout
 from ..core import *
-from ..utils import SrcLoc, get_src_loc
+from ..utils import SrcLoc, get_src_loc, MethodLayout
 
 __all__ = [
     "FIFO",
