+++ transactron/utils/gen.py
@@ -5,7 +5,8 @@ from amaranth import *
 from amaranth.back import verilog
 from amaranth.hdl import Fragment
 
-from transactron.core import TransactionManager, MethodMap, TransactionManagerKey
+from transactron.core import TransactionManager, TransactionManagerKey
+from transactron.core.manager import MethodMap
 from transactron.lib.metrics import HardwareMetricsManager
 from transactron.utils.dependencies import DependencyContext
 from transactron.utils.idgen import IdGenerator
